# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Handoff": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "acknowledgement": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledgement"
          },
          "client": {
            "title": "Client",
            "type": "string"
          },
          "content": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_applicability": {
            "title": "Current Applicability",
            "type": "string"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "execution_observed": {
            "title": "Execution Observed",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "package_state": {
            "title": "Package State",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "verification_granted": {
            "title": "Verification Granted",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "project_id",
          "run_id",
          "client",
          "digest",
          "created_at",
          "package_state",
          "current_applicability",
          "acknowledgement",
          "execution_observed",
          "verification_granted",
          "content"
        ],
        "title": "Handoff",
        "type": "object"
      },
      "HandoffList": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Handoff"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "has_more"
        ],
        "title": "HandoffList",
        "type": "object"
      },
      "HandoffRequest": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "maximum": 100000.0,
            "minimum": 512.0,
            "title": "Budget",
            "type": "integer"
          },
          "client": {
            "enum": [
              "codex",
              "claude-code",
              "copilot-vscode"
            ],
            "title": "Client",
            "type": "string"
          },
          "expected_plan": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Expected Plan",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "run_id",
          "budget",
          "expected_plan",
          "client"
        ],
        "title": "HandoffRequest",
        "type": "object"
      },
...
      "JsonValue": {},
      "Meta": {
        "properties": {
          "api_version": {
            "const": "1",
            "default": "1",
            "title": "Api Version",
            "type": "string"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "generated_at"
        ],
        "title": "Meta",
        "type": "object"
      },
...
      "Operation": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "attempts": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Attempts",
            "type": "array"
          },
          "available_at": {
            "title": "Available At",
            "type": "string"
          },
          "binding_id": {
            "default": "",
            "title": "Binding Id",
            "type": "string"
          },
          "can_cancel": {
            "title": "Can Cancel",
            "type": "boolean"
          },
          "can_retry": {
            "title": "Can Retry",
            "type": "boolean"
          },
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_id": {
            "default": "",
            "title": "Import Id",
            "type": "string"
          },
          "last_error_code": {
            "title": "Last Error Code",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "stages": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Stages",
            "type": "array"
          },
          "stages_omitted": {
            "default": 0,
            "title": "Stages Omitted",
            "type": "integer"
          },
          "state": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "action",
          "repository",
          "state",
          "revision",
          "attempt_count",
          "max_attempts",
          "created_at",
          "updated_at",
          "available_at",
          "last_error_code",
          "can_cancel",
          "can_retry",
          "stages",
          "attempts"
        ],
        "title": "Operation",
        "type": "object"
      },
      "OperationList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Operation"
            },
            "title": "Items",
            "type": "array"
          },
          "native_items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/NativeOperation"
            },
            "title": "Native Items",
            "type": "array"
          },
          "native_next_before": {
            "default": 0,
            "title": "Native Next Before",
            "type": "integer"
          },
          "next_before": {
            "title": "Next Before",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "next_before"
        ],
        "title": "OperationList",
        "type": "object"
      },
...
      "ReceiptRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Expected Digest",
            "type": "string"
          },
          "handoff_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Handoff Id",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "run_id",
          "handoff_id",
          "expected_digest"
        ],
        "title": "ReceiptRequest",
        "type": "object"
      },
...
      "Response_HandoffList_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/HandoffList"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[HandoffList]",
        "type": "object"
      },
      "Response_Handoff_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Handoff"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[Handoff]",
        "type": "object"
      },
...
      "Response_OperationList_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OperationList"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[OperationList]",
        "type": "object"
      },
      "Response_Operation_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Operation"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[Operation]",
        "type": "object"
      },
...
      "Response_TaskPreview_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskPreview"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[TaskPreview]",
        "type": "object"
      },
      "Response_Task_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Task"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[Task]",
        "type": "object"
      },
      "Response_Tasks_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Tasks"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[Tasks]",
        "type": "object"
      },
...
      "Task": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "context_status": {
            "title": "Context Status",
            "type": "string"
          },
          "current_applicability": {
            "title": "Current Applicability",
            "type": "string"
          },
          "kind": {
            "enum": [
              "external",
              "managed"
            ],
            "title": "Kind",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "kind",
          "context",
          "context_status",
          "current_applicability",
          "command"
        ],
        "title": "Task",
        "type": "object"
      },
...
      "TaskPreview": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "authority": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Authority",
            "type": "object"
          },
          "budget": {
            "title": "Budget",
            "type": "integer"
          },
          "context": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Context",
            "type": "object"
          },
          "coverage": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Coverage",
            "type": "array"
          },
          "estimated_tokens": {
            "title": "Estimated Tokens",
            "type": "integer"
          },
          "export_available": {
            "title": "Export Available",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "plan_digest": {
            "title": "Plan Digest",
            "type": "string"
          },
          "profiles": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Profiles",
            "type": "array"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "validity": {
            "items": {
              "type": "string"
            },
            "title": "Validity",
            "type": "array"
          },
          "verification_available": {
            "title": "Verification Available",
            "type": "boolean"
          }
        },
        "required": [
          "project_id",
          "run_id",
          "kind",
          "context",
          "authority",
          "plan_digest",
          "budget",
          "estimated_tokens",
          "export_available",
          "verification_available",
          "validity",
          "profiles",
          "coverage"
        ],
        "title": "TaskPreview",
        "type": "object"
      },
      "TaskVerificationRequest": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "maximum": 100000.0,
            "minimum": 512.0,
            "title": "Budget",
            "type": "integer"
          },
          "expected_plan": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Expected Plan",
            "type": "string"
          },
          "profile": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Profile",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "run_id",
          "budget",
          "expected_plan",
          "profile"
        ],
        "title": "TaskVerificationRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RepoSteward local workbench",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/commands/tasks/acknowledge": {
      "post": {
        "operationId": "acknowledgeHandoff",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,128}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Handoff_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acknowledge"
      }
    },
    "/api/v1/commands/tasks/handoff": {
      "post": {
        "operationId": "exportHandoff",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,128}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Handoff_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export"
      }
    },
    "/api/v1/commands/tasks/verify": {
      "post": {
        "operationId": "verifyTask",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,128}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Operation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify"
      }
    },
...
    "/api/v1/handoff": {
      "get": {
        "operationId": "handoff",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "handoff_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Handoff Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Handoff_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handoff"
      }
    },
    "/api/v1/handoffs": {
      "get": {
        "operationId": "handoffs",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_HandoffList_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handoffs"
      }
    },
...
    "/api/v1/operation": {
      "get": {
        "operationId": "operation",
        "parameters": [
          {
            "in": "query",
            "name": "operation_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Operation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operation"
      }
    },
    "/api/v1/operations": {
      "get": {
        "operationId": "operations",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "^(?:[0-9a-f]{32})?$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Before",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_OperationList_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operations"
      }
    },
...
    "/api/v1/task": {
      "get": {
        "operationId": "task",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Task_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task"
      }
    },
    "/api/v1/task-preview": {
      "get": {
        "operationId": "taskPreview",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "budget",
            "required": false,
            "schema": {
              "default": 24000,
              "maximum": 100000,
              "minimum": 512,
              "title": "Budget",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_TaskPreview_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview"
      }
    },
    "/api/v1/tasks": {
      "get": {
        "operationId": "tasks",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Tasks_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tasks"
      }
    },
...
  }
}